requirements.txt
dbt-postgres==1.7.10
psycopg2-binary==2.9.6
pandas==2.2.2
pytest